+++ imagingbook-common/src/main/java/imagingbook/common/threshold/global/MaxEntropyThresholder.java
@@ -37,7 +37,7 @@ public class MaxEntropyThresholder implements GlobalThresholder {
 	}
 	
 	@Override
-	public int getThreshold(int[] h) {
+	public float getThreshold(int[] h) {
 		int K = h.length;	
 		double[] p = HistogramUtils.pdf(h);			// normalized histogram (to probabilities)	
 		makeTables(p);	// initialize S0, S1
